Refactoring to Support Unit Tests
