+++ src/IssueTracker.Library/GlobalUsings.cs
@@ -2,3 +2,7 @@
 global using MongoDB.Bson;
 global using MongoDB.Bson.Serialization.Attributes;
 global using MongoDB.Driver;
+global using IssueTrackerLibrary.Contracts;
+global using IssueTrackerLibrary.Services;
+global using IssueTrackerLibrary.DataAccess;
+global using IssueTrackerLibrary.Helpers;
