+++ src/IssueTracker.UI/RegisterServices.cs
@@ -1,5 +1,6 @@
 ﻿using IssueTrackerLibrary.Contracts;
 using IssueTrackerLibrary.Helpers;
+using IssueTrackerLibrary.Services;
 
 using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 using Microsoft.Identity.Web;
